requirements.txt
aiogram>=3.0.0
aiohttp>=3.9.0
beautifulsoup4>=4.12.0
aiosqlite>=0.19.0
python-dotenv>=1.0.0
apscheduler>=3.10.0
groq>=0.4.0
feedparser>=6.0.0
lxml>=5.0.0